Add simple etag based cache
The cache only handles get requests (so in particular RelaxDB.load %w(lots of ids) is not cached), but it does handle all get requests (so views are cached).
The cache uses the etags that couchdb provides to add a If-None-Match header, if the document or view has not changed then couchdb return a 304 response and the data from the cache is returned. This means that there is not a huge gain when loading a single document (other than shifting some work off couchdb) - the cache is most useful for views. This mechanism means that there is no expiry to worry about. Couchdb knows from the etag whether the cached data is up to date or not.

The cache is off by default. To enable it pass a cache store object to RelaxDB.configure eg

RelaxDB.configure ..., :cache_store => RelaxDB::MemoryStore.new(:size => 200)

will tell RelaxDB to use a cache store that stores up to 200 entries in memory.

There is also a simple memcache based store, eg

RelaxDB.configure ..., :cache_store => RelaxDB::MemcacheStore.new('localhost:11211')

which is more useful for a web app environment. All options passed to MemcacheStore.new are passed straight through to memcache-client